Add code to exclude stacks based on an environment variable

+++ ci/add_collections.sh
@@ -8,24 +8,32 @@ then
     exit 1
 fi
 
+check_excluded_stacks () {
+    local isExcluded=0
+    for full_excluded_stack in $EXCLUDED_STACKS
+    do
+        excluded_repo=${full_excluded_stack%/*}
+        excluded_stack=${full_excluded_stack#*/}
+        if [ "$excluded_repo" == "$1" ] && [ "$excluded_stack" == "$2" ]
+        then
+            local isExcluded=1
+            break
+        fi
+    done
+    echo $isExcluded
+}
+
 base_dir="$(cd "$1" && pwd)"
 
 # directory to store assets for test or release
 assets_dir=$base_dir/ci/assets
-
-# list of repositories to build indexes for
-repo_list="experimental incubator stable"
+mkdir -p $assets_dir
 
 # url for downloading released assets
 release_url="https://github.com/$TRAVIS_REPO_SLUG/releases/download"
 
-# dockerhub org for publishing stack
-export DOCKERHUB_ORG=appsody
-
-mkdir -p $assets_dir
-
 # iterate over each repo
-for repo_name in $repo_list
+for repo_name in $REPO_LIST
 do
     repo_dir=$base_dir/$repo_name
     if [ -d $repo_dir ]
@@ -53,39 +61,45 @@ do
             then
                 i=0
                 stack_id=$(basename $stack_dir)
-                stack_version=$(awk '/^version *:/ { gsub("version:","",$NF); gsub("\"","",$NF); print $NF}' $stack)
-                collection=$stack_dir/collection.yaml
+                excluded=0
+                if [ -n "$EXCLUDED_STACKS" ]; then 
+                    excluded=$( check_excluded_stacks $repo_name $stack_id )
+                fi
+                if [ $excluded -eq 0 ]; then
+                    stack_version=$(awk '/^version *:/ { gsub("version:","",$NF); gsub("\"","",$NF); print $NF}' $stack)
+                    collection=$stack_dir/collection.yaml
 
-                count=0
-                stack_to_use=-1
-                while [ $count -lt $num_stacks ]
-                do
-                    if [ $stack_id == $(yq r $all_stacks stacks.[$count].id) ]
+                    count=0
+                    stack_to_use=-1
+                    while [ $count -lt $num_stacks ]
+                    do
+                        if [ $stack_id == $(yq r $all_stacks stacks.[$count].id) ]
+                        then
+                            stack_to_use=$count
+                            break;
+                        fi
+                        count=$(( $count + 1 ))
+                    done
+                    if [ $stack_to_use -ge 0 ]
                     then
-                        stack_to_use=$count
-                        count=9999
+                        yq r $all_stacks stacks.[$stack_to_use] > $one_stack
+                        if [ -f $collection ]
+                        then
+                            if [ -f $base_dir/ci/add_collection_resources.sh ]
+                            then
+                                . $base_dir/ci/add_collection_resources.sh $base_dir $stack_dir $stack_version $repo_name $one_stack
+                            fi
+                        fi
+                        yq p -i $one_stack stacks.[+]
+                        yq m -a -i $index_file_v2_temp $one_stack
                     fi
-                    count=$(( $count + 1 ))
-                done
-                if [ $stack_to_use -ge 0 ]
-                then
-                    #yq r $all_stacks stacks.[$stack_to_use] | yq p - stacks.[+] > $one_stack
-                    yq r $all_stacks stacks.[$stack_to_use] > $one_stack
-                fi
-                
-                if [ -f $collection ]
-                then
-                    if [ -f $base_dir/ci/add_collection_resources.sh ]
+                    if [ -f $one_stack ]
                     then
-                       . $base_dir/ci/add_collection_resources.sh $base_dir $stack_dir $stack_version $repo_name $one_stack
+                        rm -f $one_stack
                     fi
+                else
+                   echo "--- The $stack_id collection has been excluded from the build"
                 fi
-                yq p -i $one_stack stacks.[+]
-                yq m -a -i $index_file_v2_temp $one_stack
-            fi
-            if [ -f $one_stack ]
-            then
-                rm -f $one_stack
             fi
         done
         mv $index_file_v2_temp $index_file_v2
